MessagingHistory/requirements.txt
langchain==0.3.18
langchain-community== 0.3.17
langchain-core==0.3.35
langchain-google-genai==2.0.9
langchain-text-splitters==0.3.6
google-genai==1.2.0
python-dotenv==1.0.1
fastapi==0.115.8
langserve==0.3.1
sse_starlette==2.2.1
langchain-chroma==0.2.2